+++ hutool-core/src/main/java/cn/hutool/core/util/BooleanUtil.java
@@ -380,7 +380,7 @@ public class BooleanUtil {
 		}
 
 		for (final Boolean b : array) {
-			if(isFalse(b)){
+			if(!isTrue(b)){
 				return false;
 			}
 		}